{{cookiecutter.repo_name}}/requirements/base.txt
tornado>=5.0.0
evision-lib>=0.1.2
requests>=2.20.1
